Move HeatmapLayer files
