+++ src/eventLists/HeatmapLayerEventList.js
@@ -1,4 +1,4 @@
-// https://developers.google.com/maps/documentation/javascript/3.exp/reference#Polyline
+// https://developers.google.com/maps/documentation/javascript/3.exp/reference#HeatmapLayer
 // [].map.call($0.querySelectorAll("tr>td>code"), function(it){ return it.textContent; })
 export default [
   `zoom_changed`,
